2.38.2-gconvert-cygwin.patch
POSIX is clear that a path starting with '//' has a system-dependent
meaning, and on Cygwin, is interpreted as an UNC path.  Never
accidentally produce those when converting a URI to filename.

Fix handling of file: URIs starting with three or four slashes. Convert
these to filenames starting with '///', which is never a UNC path.

Also adjust tests to match change to g_filename_from_uri.